--- a/include/seqan3/search/fm_index/bi_fm_index_cursor.hpp
+++ b/include/seqan3/search/fm_index/bi_fm_index_cursor.hpp
@@ -1,6 +1,6 @@
 // -----------------------------------------------------------------------------------------------------
-// Copyright (c) 2006-2019, Knut Reinert & Freie Universität Berlin 
-// Copyright (c) 2016-2019, Knut Reinert & MPI für molekulare Genetik 
+// Copyright (c) 2006-2019, Knut Reinert & Freie Universität Berlin
+// Copyright (c) 2016-2019, Knut Reinert & MPI für molekulare Genetik
 // This file may be used, modified and/or redistributed under the terms of the 3-clause BSD-License
 // shipped with this file and also available at: https://github.com/seqan/seqan3/blob/master/LICENSE
 // -----------------------------------------------------------------------------------------------------
@@ -42,12 +42,12 @@ namespace seqan3
  * the operation was successful or not. In case of an unsuccessful operation the cursor remains unmodified, i.e. an
  * cursor can never be in an invalid state except default constructed cursors that are always invalid.
  *
- * \cond DEV
+ * \if DEV
  *     The behaviour is equivalent to a prefix and suffix tree with the space and time efficiency of the underlying pure
  *     FM indices. The cursor traverses the implicit prefix and suffix trees beginning at the root node. The implicit
  *     prefix and suffix trees are not compacted, i.e. going down an edge using extend_right(char) will increase the
  *     query by only one character.
- * \endcond
+ * \endif
  *
  * The asymptotic running times for using the cursor depend on the SDSL index configuration. To determine the exact
  * running times, you have to additionally look up the running times of the used traits (configuration).
@@ -82,6 +82,8 @@ protected:
 
     //!\brief Type of the representation of characters in the underlying SDSL index.
     using sdsl_char_type = typename index_type::sdsl_char_type;
+    //!\brief Type of the alphabet size in the underlying SDSL index.
+    using sdsl_sigma_type = typename index_type::sdsl_sigma_type;
 
     //!\brief Type of the underlying FM index.
     index_type const * index;
@@ -98,6 +100,12 @@ protected:
     //!\brief Right suffix array interval of the reverse cursor (for extend_left).
     size_type rev_rb;
     //\}
+
+    //!\brief Alphabet size of the index without delimiters
+    sdsl_sigma_type sigma;
+
+    //!\brief Indicates whether index is built over a collection
+    static bool constexpr is_collection = dimension_v<typename index_type::text_type> == 2;
 
     /*!\name Information for on cycle_back() and cycle_front()
      * \brief Only stored for the cursor that has been used last to go down an edge because once one cursor is
@@ -243,9 +251,10 @@ public:
     bi_fm_index_cursor & operator=(bi_fm_index_cursor &&) noexcept = default;
 
     bi_fm_index_cursor(index_t const & _index) noexcept : index(&_index),
-                                                           fwd_lb(0), fwd_rb(_index.size() - 1),
-                                                           rev_lb(0), rev_rb(_index.size() - 1),
-                                                           depth(0)
+                                                          fwd_lb(0), fwd_rb(_index.size() - 1),
+                                                          rev_lb(0), rev_rb(_index.size() - 1),
+                                                          sigma(_index.fwd_fm.index.sigma - is_collection),
+                                                          depth(0)
     {}
     //\}
 
@@ -293,9 +302,9 @@ public:
 
     /*!\brief Tries to extend the query by the smallest possible character to the right such that the query is found in
      *        the text.
-     *        \cond DEV
+     *        \if DEV
      *            Goes down the leftmost (i.e. lexicographically smallest) edge.
-     *        \endcond
+     *        \endif
      * \returns `true` if the cursor could extend the query successfully.
      *
      * ### Complexity
@@ -319,14 +328,14 @@ public:
         size_type new_parent_lb = fwd_lb, new_parent_rb = fwd_rb;
 
         sdsl_char_type c = 1; // NOTE: start with 0 or 1 depending on implicit_sentintel
-        while (c < index->fwd_fm.index.sigma &&
+        while (c < sigma &&
                !bidirectional_search(index->fwd_fm.index, index->fwd_fm.index.comp2char[c],
                                      fwd_lb, fwd_rb, rev_lb, rev_rb))
         {
             ++c;
         }
 
-        if (c != index->fwd_fm.index.sigma)
+        if (c != sigma)
         {
             parent_lb = new_parent_lb;
             parent_rb = new_parent_rb;
@@ -341,9 +350,9 @@ public:
 
     /*!\brief Tries to extend the query by the smallest possible character to the left such that the query is found in
      *        the text.
-     *        \cond DEV
+     *        \if DEV
      *            Goes down the leftmost (i.e. lexicographically smallest) edge in the reverse cursor.
-     *        \endcond
+     *        \endif
      * \returns `true` if the cursor could extend the query successfully.
      *
      * ### Complexity
@@ -367,14 +376,14 @@ public:
         size_type new_parent_lb = rev_lb, new_parent_rb = rev_rb;
 
         sdsl_char_type c = 1; // NOTE: start with 0 or 1 depending on implicit_sentintel
-        while (c < index->rev_fm.index.sigma &&
+        while (c < sigma &&
                !bidirectional_search(index->rev_fm.index, index->rev_fm.index.comp2char[c],
                                      rev_lb, rev_rb, fwd_lb, fwd_rb))
         {
             ++c;
         }
 
-        if (c != index->rev_fm.index.sigma)
+        if (c != sigma)
         {
             parent_lb = new_parent_lb;
             parent_rb = new_parent_rb;
@@ -597,12 +606,12 @@ public:
 
     /*!\brief Tries to replace the rightmost character of the query by the next lexicographically larger character such
      *        that the query is found in the text.
-     *        \cond DEV
+     *        \if DEV
      *            Moves the cursor to the right sibling of the current suffix tree node. It would be equivalent to
      *            going up an edge and going down that edge with the smallest character that is larger than the
      *            previous searched character. Calling cycle_*() on an cursor pointing to the root node is undefined
      *            behaviour!
-     *        \endcond
+     *        \endif
      * \returns `true` if there exists a query in the text where the rightmost character of the query is
      *          lexicographically larger than the current rightmost character of the query.
      *
@@ -632,14 +641,14 @@ public:
 
         sdsl_char_type c = _last_char + 1;
 
-        while (c < index->fwd_fm.index.sigma &&
+        while (c < sigma &&
                !bidirectional_search_cycle(index->fwd_fm.index, index->fwd_fm.index.comp2char[c],
                                            parent_lb, parent_rb, fwd_lb, fwd_rb, rev_lb, rev_rb))
         {
             ++c;
         }
 
-        if (c != index->fwd_fm.index.sigma)
+        if (c != sigma)
         {
             _last_char = c;
 
@@ -650,12 +659,12 @@ public:
 
     /*!\brief Tries to replace the leftmost character of the query by the next lexicographically larger character such
      *        that the query is found in the text.
-     *        \cond DEV
+     *        \if DEV
      *            Moves the cursor to the right sibling of the current suffix tree node. It would be equivalent to
      *            going up an edge and going down that edge with the smallest character that is larger than the
      *            previous searched character. Calling cycle_*() on an cursor pointing to the root node is undefined
      *            behaviour!
-     *        \endcond
+     *        \endif
      * \returns `true` if there exists a query in the text where the leftmost character of the query is
      *          lexicographically larger than the current leftmost character of the query.
      *
@@ -684,14 +693,14 @@ public:
         assert(index != nullptr && query_length() > 0);
 
         sdsl_char_type c = _last_char + 1;
-        while (c < index->rev_fm.index.sigma &&
+        while (c < sigma &&
                !bidirectional_search_cycle(index->rev_fm.index, index->rev_fm.index.comp2char[c],
                                            parent_lb, parent_rb, rev_lb, rev_rb, fwd_lb, fwd_rb))
         {
             ++c;
         }
 
-        if (c != index->rev_fm.index.sigma)
+        if (c != sigma)
         {
             _last_char = c;
 
@@ -804,6 +813,9 @@ public:
      *
      * \snippet test/snippet/search/bi_fm_index_cursor.cpp to_rev_cursor
      *
+     * \attention When the index is built for text collections, the returned text IDs will be reversed.
+     *
+     * \snippet test/snippet/search/bi_fm_index_cursor.cpp to_rev_cursor_collection
      * ### Complexity
      *
      * Constant.
@@ -834,9 +846,9 @@ public:
     }
 
     /*!\brief Returns the searched query.
-     *        \cond DEV
+     *        \if DEV
      *            Returns the concatenation of all edges from the root node to the cursors current node.
-     *        \endcond
+     *        \endif
      *
      * ### Complexity
      *
@@ -847,11 +859,27 @@ public:
      * No-throw guarantee.
      */
     auto query() const noexcept
+    //!\cond
+        requires !is_collection
+    //!\endcond
     {
         assert(index != nullptr && index->text != nullptr);
 
         size_type const query_begin = offset() - index->fwd_fm.index[fwd_lb];
         return *index->text | ranges::view::slice(query_begin, query_begin + query_length());
+    }
+
+    //!\overload
+    auto query() const noexcept
+    //!\cond
+        requires is_collection
+    //!\endcond
+    {
+        assert(index != nullptr && index->text != nullptr);
+
+        size_type const loc = offset() - index->fwd_fm.index[fwd_lb];
+        size_type const query_begin = loc - index->fwd_fm.text_begin_rs.rank(loc + 1) + 1; // Substract delimiters
+        return *index->text | ranges::view::join | ranges::view::slice(query_begin, query_begin + query_length());
     }
 
     //!\copydoc query()
@@ -892,6 +920,9 @@ public:
      * Strong exception guarantee (no data is modified in case an exception is thrown).
      */
     std::vector<size_type> locate() const
+    //!\cond
+        requires !is_collection
+    //!\endcond
     {
         assert(index != nullptr);
 
@@ -899,6 +930,26 @@ public:
         for (size_type i = 0; i < occ.size(); ++i)
         {
             occ[i] = offset() - index->fwd_fm.index[fwd_lb + i];
+        }
+        return occ;
+    }
+
+    //!\overload
+    std::vector<std::pair<size_type, size_type>> locate() const
+    //!\cond
+        requires is_collection
+    //!\endcond
+    {
+        assert(index != nullptr);
+
+        std::vector<std::pair<size_type, size_type>> occ;
+        occ.reserve(count());
+        for (size_type i = 0; i < count(); ++i)
+        {
+            size_type loc = offset() - index->fwd_fm.index[fwd_lb + i];
+            size_type sequence_rank = index->fwd_fm.text_begin_rs.rank(loc + 1);
+            size_type sequence_position = loc - index->fwd_fm.text_begin_ss.select(sequence_rank);
+            occ.emplace_back(sequence_rank - 1, sequence_position);
         }
         return occ;
     }
@@ -916,6 +967,9 @@ public:
      * Strong exception guarantee (no data is modified in case an exception is thrown).
      */
     auto lazy_locate() const
+    //!\cond
+        requires !is_collection
+    //!\endcond
     {
         assert(index != nullptr);
 
@@ -923,6 +977,27 @@ public:
              | view::transform([*this, _offset = offset()] (auto sa_pos)
                {
                    return _offset - index->fwd_fm.index[sa_pos];
+               });
+    }
+
+    //!\overload
+    auto lazy_locate() const
+    //!\cond
+        requires is_collection
+    //!\endcond
+    {
+        assert(index != nullptr);
+
+        return ranges::view::iota(fwd_lb, fwd_lb + count())
+               | view::transform([*this, _offset = offset()] (auto sa_pos)
+               {
+                   return _offset - index->fwd_fm.index[sa_pos];
+               })
+               | view::transform([*this] (auto loc)
+               {
+                   size_type sequence_rank = index->fwd_fm.text_begin_rs.rank(loc + 1);
+                   size_type sequence_position = loc - index->fwd_fm.text_begin_ss.select(sequence_rank);
+                   return std::make_pair(sequence_rank-1, sequence_position);
                });
     }