--- a/test/unit/search/CMakeLists.txt
+++ b/test/unit/search/CMakeLists.txt
@@ -1,9 +1,6 @@
 add_subdirectories()
 
 seqan3_test (sdsl_index_test.cpp)
-seqan3_test (fm_index_test.cpp)
-seqan3_test (fm_index_cursor_test.cpp)
-seqan3_test (bi_fm_index_cursor_test.cpp)
 
 seqan3_test (search_configuration_test.cpp)
 seqan3_test (search_scheme_test.cpp)