… include/seqan3/search/fm_index/fm_index.hpp
… test/snippet/search/bi_fm_index_cursor.cpp
… test/snippet/search/fm_index_collection.cpp
… test/unit/search/bi_fm_index_cursor_test.cpp
… test/unit/search/fm_index/CMakeLists.txt
… test/unit/search/fm_index/bi_fm_index_aa27_test.cpp
… test/unit/search/fm_index/bi_fm_index_char_test.cpp
… test/unit/search/fm_index/bi_fm_index_dna4_test.cpp
… test/unit/search/fm_index/fm_index_collection_test_template.hpp
… test/unit/search/fm_index/fm_index_dna4_test.cpp
… test/unit/search/fm_index/fm_index_test_template.hpp
… test/unit/search/fm_index_cursor/CMakeLists.txt
… test/unit/search/fm_index_cursor/bi_fm_index_cursor_collection_test.cpp
… test/unit/search/fm_index_cursor/bi_fm_index_cursor_collection_test_template.hpp
… test/unit/search/fm_index_cursor/bi_fm_index_cursor_test.cpp
… test/unit/search/fm_index_cursor/bi_fm_index_cursor_test_template.hpp
… test/unit/search/fm_index_cursor/fm_index_cursor_collection_test.cpp
… test/unit/search/fm_index_cursor/fm_index_cursor_collection_test_template.hpp
… test/unit/search/fm_index_cursor/fm_index_cursor_test.cpp
… test/unit/search/fm_index_cursor/fm_index_cursor_test_template.hpp
… test/unit/search/fm_index_cursor_test.cpp
… test/unit/search/helper.hpp