Resolves #602

**Main problem**
`bi_fm_index_iterator_collection_test` uses too much memory: ~9GB for me, Travis-VM only has 7.5GB.
<details><summary>Error message (click me)</summary><p>

```
[ 96%] Building CXX object search/fm_index_iterator/CMakeFiles/bi_fm_index_iterator_collection_coverage_test.dir/bi_fm_index_iterator_collection_test.cpp.o
virtual memory exhausted: Cannot allocate memory
search/fm_index_iterator/CMakeFiles/bi_fm_index_iterator_collection_coverage_test.dir/build.make:62: recipe for target 'search/fm_index_iterator/CMakeFiles/bi_fm_index_iterator_collection_coverage_test.dir/bi_fm_index_iterator_collection_test.cpp.o' failed
make[2]: *** [search/fm_index_iterator/CMakeFiles/bi_fm_index_iterator_collection_coverage_test.dir/bi_fm_index_iterator_collection_test.cpp.o] Error 1
make[2]: Target 'search/fm_index_iterator/CMakeFiles/bi_fm_index_iterator_collection_coverage_test.dir/build' not remade because of errors.
CMakeFiles/Makefile2:5969: recipe for target 'search/fm_index_iterator/CMakeFiles/bi_fm_index_iterator_collection_coverage_test.dir/all' failed
make[1]: *** [search/fm_index_iterator/CMakeFiles/bi_fm_index_iterator_collection_coverage_test.dir/all] Error 2
```
</p></details>
<br>

The test only uses `bi_fm_index_iterator<bi_fm_index<std::vector<std::vector<dna4>>>>`
Same problem for `fm_index_iterator_colletion_test` that tests the common functionality of `fm_index_iterator` and `bi_fm_index_iterator` (latter is currently commented out)
<details><summary>Some time -v and -fmem-report stuff (click me)</summary><p>

```
[ 75%] Building CXX object search/fm_index_iterator/CMakeFiles/bi_fm_index_iterator_collection_test.dir/bi_fm_index_iterator_collection_test.cpp.o
Number of expanded macros:                     24228
Average number of tokens per macro expansion:      8

Line Table allocations during the compilation process
Number of ordinary maps used:         1949 
Ordinary map used size:                 60k
Number of ordinary maps allocated:    4096 
Ordinary maps allocated size:          128k
Number of macro maps used:              21k
Macro maps used size:                  874k
Macro maps locations size:            1656k
Macro maps size:                      2531k
Duplicated maps locations size:        601k
Total allocated maps size:            2808k
Total used maps size:                 2591k
Ad-hoc table size:                      12M
Ad-hoc table entries used:           354735
optimized_ranges: 1226617
unoptimized_ranges: 326453

Memory still allocated at the end of the compilation process
Size      Allocated        Used    Overhead
8               684k        227k         20k
16             8564k       5577k        183k
32              137M         75M       2470k
64               39M         25M        638k
256              99M         92M       1388k
512            9060k       6227k        123k
1024             50M         37M        707k
2048             78M         67M       1099k
4096            134M        134M       1887k
8192            265M        265M       1861k
16384           292M        292M       1022k
32768           444M        444M        777k
65536           738M        738M        646k
131072         1042M       1042M        456k
262144          996M        996M        218k
524288          407M        407M         44k
1048576          22M         22M       1232 
2097152        2048k       2048k         56 
4194304        8192k       8192k        112 
8388608        8192k       8192k         56 
16777216         16M         16M         56 
24               75M         40M       1351k
40              172M        102M       2765k
48              186M         28M       2981k
56             9836k       3491k        153k
72               11M       3321k        166k
80               92M         92M       1301k
88               10M       5762k        140k
96               26M         23M        376k
112            7556k       2734k        103k
120            4368k       2232k         59k
152             242M        230M       3400k
128              84M         78M       1182k
144              25M       8527k        355k
168             590M        518M       8265k
176            2972k       2018k         40k
104              32M         28M        461k
368              37M         36M        527k
264             800k        441k         10k
Total          6415M       5892M         36M

String pool
entries		72150
identifiers	61302 (84.96%)
slots		131072
deleted		10848
bytes		8155k (17592186044408M overhead)
table size	1024k
coll/search	0.6983
ins/search	0.0721
avg. entry	115.75 bytes (+/- 194.90)
longest entry	1528
(No per-node statistics)
Type hash: size 262139, 176613 elements, 0.829618 collisions
DECL_DEBUG_EXPR  hash: size 1021, 0 elements, 0.000000 collisions
DECL_VALUE_EXPR  hash: size 4093, 314 elements, 0.992443 collisions
no search statistics
decl_specializations: size 1048573, 642941 elements, 1.817459 collisions
type_specializations: size 524287, 327955 elements, 1.682568 collisions
No gimple statistics
No RTX statistics

Heap vectors                                            Leak           Peak     Times       Leak items Peak items
--------------------------------------------------------------------------------------------------------------------------------------------
--------------------------------------------------------------------------------------------------------------------------------------------
Total                                                      0                        0                0
--------------------------------------------------------------------------------------------------------------------------------------------


Alias oracle query stats:
  refs_may_alias_p: 0 disambiguations, 0 queries
  ref_maybe_used_by_call_p: 0 disambiguations, 0 queries
  call_may_clobber_ref_p: 0 disambiguations, 0 queries
  TBAA oracle: 0 disambiguations 5 queries
               0 are in alias set 0
               0 queries asked about the same object
               5 queries asked about the same alias set
               0 access volatile
               0 are dependent in the DAG
               0 are aritificially in conflict with void *

PTA query stats:
  pt_solution_includes: 0 disambiguations, 2046 queries
  pt_solutions_intersect: 0 disambiguations, 0 queries
[100%] Linking CXX executable bi_fm_index_iterator_collection_test
[100%] Built target bi_fm_index_iterator_collection_test
	Command being timed: "make bi_fm_index_iterator_collection_test"
	User time (seconds): 524.51
	System time (seconds): 24.06
	Percent of CPU this job got: 94%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 9:41.67
	Average shared text size (kbytes): 0
	Average unshared data size (kbytes): 0
	Average stack size (kbytes): 0
	Average total size (kbytes): 0
	Maximum resident set size (kbytes): 8615476
	Average resident set size (kbytes): 0
	Major (requiring I/O) page faults: 8279
	Minor (reclaiming a frame) page faults: 4772575
	Voluntary context switches: 13195
	Involuntary context switches: 677196
	Swaps: 0
	File system inputs: 0
	File system outputs: 0
	Socket messages sent: 0
	Socket messages received: 0
	Signals delivered: 84
	Page size (bytes): 4096
	Exit status: 0
```

</p></details>
<br>

- [X] Snippets
- [x] Make Debug tests work on travis (coverage tests)
- - After above: Check if there's redundant code in test cases that can be refactored
- [x] Docs
- [X] FM
- [X] Bi-FM
- [x] Clean up code
- [x] Put debug stream for pairs into separate PR

<details><summary>Old questions (click me)</summary><p>

<details><summary>Regarding the output order (click me)</summary><p>

```cpp
int main()
{
    std::vector<std::vector<dna4>> vec { "ACAGCGAGC"_dna4, "GTTTCGAT"_dna4,
                                         "GCGATCGTAGCT"_dna4, "CACGCG"_dna4};
    fm_index index{vec};
    auto it = index.begin();
    it.extend_right("GCG"_dna4);
    debug_stream << it.query() << '\n';
    debug_stream << it.last_char() << '\n';
    for (auto && x : it.locate())
        debug_stream << x.first << '\t' << x.second << '\n';
}
```

Would print:
```
GCG    
G
0    3
3    3
2    0
```

Which is all correct, but the <sequence_no, sequence_pos> pairs are not ordered by sequence_no since the order in the index does not correspond with the order of the texts in the collection.

Should it be ordered?

**Answer**: It's fine but should be documented.

</p></details>
<br>

<details><summary>Copying the iterator multiple times in lazy_locate() (click me)</summary><p>

```cpp
return ranges::view::iota(node.lb, node.lb + count())
               | view::transform([*this, _offset = offset()] (auto sa_pos) { return _offset - index->index[sa_pos]; })
               | view::filter([*this, _query_length = query_length()] (auto loc)
               {
                   return (index->text_begin_rs.rank(loc + 1) == index->text_begin_rs.rank(loc + _query_length));
               })
               | view::transform([*this] (auto loc)
               {
                   size_type sequence_rank = index->text_begin_rs.rank(loc + 1);
                   size_type sequence_position = loc - index->text_begin_ss.select(sequence_rank);
                   return std::make_pair(sequence_rank-1, sequence_position);
               });
```

Do I need to copy the fm_index_iterator multiple times?

```cpp
| view::transform([*this, _offset = offset()] 
| view::filter([*this, _query_length = query_length()]
| view::transform([*this] 
```

**Answer**: It's fine since there's only stack memory involved when copying.

</p></details>
<br>
<details><summary>Boundary check (click me)</summary><p>

Currently, we are checking the validity of the match when calling `locate()` or `lazy_locate()`.
Actually, we should check for it, when moving in the index (e.g. `extend_right()`).
`locate()` will not always be called, so incorrect iterators might be used downstream.
But since the SA range is contiguous, it's hard to modify it to exclude matches crossing the boundaries that may be in the middle of the SA range.

**Answer**: Resolved by using delimiters.

</p></details>
<br>
</p></details>
<br>