@marehr 2018-11-28
Tests are missing right now, for the collections extension.