@eseiler 2018-11-29
Question:

The (bi)_fm_index_concept won't hold for collections:

```cpp
requires requires (t index, std::vector<typename t::char_type> const text)
    {
        { t(text) };
        { index.construct(text) } -> void;
    };
```

with `using char_type = innermost_value_type_t<text_t>;`.
So it will resolve, e.g. `std::vector<std::vector<dna4>>` to `dna4` and then try to construct an index with `std::vector<dna4>` which a fm_index for `std::vector<std::vector<dna4>>` cannot do.

So, either:

1) Change 
```cpp
using char_type = innermost_value_type_t<text_t>;
```
to
```cpp
using char_type = value_type_t<text_t>;
```

2) Change 
```cpp
requires requires (t index, std::vector<typename t::char_type> const text)
``` 
to 
```cpp
requires requires (t index, std::vector<value_type_t<typename t::text_type>> const text)
```

3) Change 
```cpp
requires requires (t index, std::vector<typename t::char_type> const text)
``` 
to 
```cpp
requires requires (t index, typename t::text_type const text)
```