@eseiler 2018-11-30
I transferred the debug_stream changes to #596