@eseiler 2018-12-03
I added changes to use a delimiter for the texts.

➕iterator now won't cross the boundaries
➖increases alphabet size by 1 
➖char alphabet currently not indexable (256 + 1 > 8 bit unsigned int)

I also managed to reduce the number of text copies to 1 (only the copy text -> sdsl data type is done now).