@eseiler 2018-12-07
I'll have a look at it once we can actually build the tests in debug mode.