@eseiler 2018-12-10
> > Regarding the output order (click me)
> 
> Not sorting is the way to go I think. If people want it sorted, they can always sort it themselves. But it should be documented clearly.

👍 

> 
> > I added changes to use a delimiter for the texts.
> 
> This looks like the most reasonable thing to do, but will this affect the performance of the searches on wavelet-trees / depth of the wt? And do you know what the 0-rank is used for by the SDSL? Could you maybe hijack that?
0 is used as sentinel. Using delimiters should affect the performance since we add another symbol to the alphabet.
> 
> For `char` you could scan the input for a character that is not used and use that as delimiter (and throw if all of the char characters are used).

Currently I'm reserving rank 255 for char. If the char is signed, 255 will never be used anyway, for unsigned chars it's `ÿ` 

> > Copying the iterator multiple times in lazy_locate() (click me)
> 
> If you use delimiters, do you still need to filter? in general it is not a problem to copy the iterator/cursor, because it has only stack storage AFAIK.

No I don't filter there anymore, it's an old snippet. 

> > Boundary check (click me)
> 
> This is solved by using delimiters, right?

Yes