@eseiler 2018-12-17
Let's see if this works.
I adapted the view for fm-indices over collections to copy once
https://github.com/seqan/seqan3/blob/767c2d60be70932c8bf01c201b534c463e538f5d/include/seqan3/search/fm_index/fm_index.hpp#L346-L362
And it takes considerably less memory to build it in debug mode.