@eseiler 2018-12-19
From my POV this is ready to be merged.
For the additional stuff I want to have a look at, I already created cards.